ch: separation of concerns
